backend/requirements.txt
admet_ai==2.0.1
aimsim_core==2.2.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
astartes==1.3.3
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
backcall==0.2.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==26.3.1
bleach==6.3.0
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
chemprop==2.2.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ConfigArgParse==1.7.5
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cycler==0.12.1
daphne==4.2.1
decorator==5.2.1
defusedxml==0.7.1
descriptastorus==2.8.0
dill==0.4.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
docstring_parser==0.17.0
drf-spectacular==0.29.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
flake8==7.3.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.5.0
kombu==5.6.2
lightning==2.6.1
lightning-utilities==0.15.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mhfp==1.9.6
mistune==3.2.0
mordredcommunity==2.0.7
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
padelpy==0.1.16
pandas==3.0.1
pandas-flavor==0.8.1
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.1
pipreqs==0.5.0
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
pytokens==0.4.1
pytorch-lightning==2.6.1
PyYAML==6.0.3
pyzmq==27.1.0
rdkit==2025.9.6
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.8
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.10.0
torchmetrics==1.9.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
Twisted==25.5.0
txaio==25.12.2
typed-argument-parser==1.11.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
xarray==2026.2.0
yarg==0.1.9
yarl==1.23.0
zope.interface==8.2
